all permisisons checks complete

+++ structures/guild.ts
@@ -20,6 +20,8 @@ import { create_channel } from "./channel.ts"
 import { Channel_Create_Options } from "../types/channel.ts"
 import { Image_Size, Image_Formats } from "../types/cdn.ts"
 import { Permissions, Permission } from "../types/permission.ts"
+import { bot_has_permission } from "../utils/permissions.ts"
+import { Errors } from "../types/errors.ts"
 
 export const create_guild = (data: Create_Guild_Payload, client: Client) => {
   const guild = {
@@ -98,7 +100,8 @@ export const create_guild = (data: Create_Guild_Payload, client: Client) => {
       data.banner ? format_image_url(endpoints.GUILD_BANNER(data.id, data.banner), size, format) : undefined,
     /** Create a channel in your server. Bot needs MANAGE_CHANNEL permissions in the server. */
     create_channel: (name: string, options: Channel_Create_Options) => {
-      // TODO: Check if the bot has `MANAGE_CHANNELS` permission before making a channel
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_CHANNELS]))
+        throw new Error(Errors.MISSING_MANAGE_CHANNELS)
       return client.discordRequestManager.post(endpoints.GUILD_CHANNELS(data.id), {
         name,
         type: options?.type ? ChannelTypes[options.type] : undefined,
@@ -135,7 +138,8 @@ export const create_guild = (data: Create_Guild_Payload, client: Client) => {
     },
     /** Create an emoji in the server. Emojis and animated emojis have a maximum file size of 256kb. Attempting to upload an emoji larger than this limit will fail and return 400 Bad Request and an error message, but not a JSON status code. */
     create_emoji: (name: string, image: string, options: Create_Emojis_Options) => {
-      // TODO: Check if the bot has `MANAGE_EMOJIS` permission
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_EMOJIS]))
+        throw new Error(Errors.MISSING_MANAGE_EMOJIS)
       return client.discordRequestManager.post(endpoints.GUILD_EMOJIS(data.id), {
         ...options,
         name,
@@ -144,7 +148,8 @@ export const create_guild = (data: Create_Guild_Payload, client: Client) => {
     },
     /** Modify the given emoji. Requires the MANAGE_EMOJIS permission. */
     edit_emoji: (id: string, options: Edit_Emojis_Options) => {
-      // TODO: check if the bot has `MANAGE_EMOJIS` permission
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_EMOJIS]))
+        throw new Error(Errors.MISSING_MANAGE_EMOJIS)
       return client.discordRequestManager.patch(endpoints.GUILD_EMOJI(data.id, id), {
         name: options.name,
         roles: options.roles
@@ -152,12 +157,14 @@ export const create_guild = (data: Create_Guild_Payload, client: Client) => {
     },
     /** Delete the given emoji. Requires the MANAGE_EMOJIS permission. Returns 204 No Content on success. */
     delete_emoji: (id: string, reason?: string) => {
-      // TODO: check if the bot has `MANAGE_EMOJIS` permission
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_EMOJIS]))
+        throw new Error(Errors.MISSING_MANAGE_EMOJIS)
       return client.discordRequestManager.delete(endpoints.GUILD_EMOJI(data.id, id), { reason })
     },
     /** Create a new role for the guild. Requires the MANAGE_ROLES permission. */
     create_role: async (options: Create_Role_Options) => {
-      // TODO: check if the bot has the `MANAGE_ROLES` permission.
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_ROLES]))
+        throw new Error(Errors.MISSING_MANAGE_ROLES)
       const role = await client.discordRequestManager.post(endpoints.GUILD_ROLES(data.id), {
         ...options,
         permissions: options.permissions?.map(perm => Permissions[perm])
@@ -168,12 +175,14 @@ export const create_guild = (data: Create_Guild_Payload, client: Client) => {
     },
     /** Edit a guild role. Requires the MANAGE_ROLES permission. */
     edit_role: (id: string, options: Create_Role_Options) => {
-      // TODO: check if the bot has the `MANAGE_ROLES` permission.
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_ROLES]))
+        throw new Error(Errors.MISSING_MANAGE_ROLES)
       return client.discordRequestManager.patch(endpoints.GUILD_ROLE(data.id, id), options)
     },
     /** Delete a guild role. Requires the MANAGE_ROLES permission. */
     delete_role: (id: string) => {
-      // TODO: check if the bot has the `MANAGE_ROLES` permission.
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_ROLES]))
+        throw new Error(Errors.MISSING_MANAGE_ROLES)
       return client.discordRequestManager.delete(endpoints.GUILD_ROLE(data.id, id))
     },
     /** Returns a list of role objects for the guild.
@@ -181,27 +190,29 @@ export const create_guild = (data: Create_Guild_Payload, client: Client) => {
      * ⚠️ **If you need this, you are probably doing something wrong. This is not intended for use. Your roles will be cached in your guild.**
      */
     get_roles: () => {
-      // TODO: check if the bot has the `MANAGE_ROLES` permission.
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_ROLES]))
+        throw new Error(Errors.MISSING_MANAGE_ROLES)
       return client.discordRequestManager.get(endpoints.GUILD_ROLES(data.id))
     },
     /** Modify the positions of a set of role objects for the guild. Requires the MANAGE_ROLES permission. */
     swap_roles: (rolePositons: Position_Swap) => {
-      // TODO: check if the bot has the `MANAGE_ROLES` permission.
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_ROLES]))
+        throw new Error(Errors.MISSING_MANAGE_ROLES)
       return client.discordRequestManager.patch(endpoints.GUILD_ROLES(data.id), rolePositons)
     },
     /** Check how many members would be removed from the server in a prune operation. Requires the KICK_MEMBERS permission */
     get_prune_count: async (days: number) => {
-      if (days < 1) {
-        throw `The number of days to count prune for must be 1 or more.`
-      }
-      // TODO: check if the bot has `KICK_MEMBERS` permission
+      if (days < 1) throw new Error(Errors.PRUNE_MIN_DAYS)
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.KICK_MEMBERS]))
+        throw new Error(Errors.MISSING_KICK_MEMBERS)
       const result = (await client.discordRequestManager.get(endpoints.GUILD_PRUNE(data.id), { days })) as PrunePayload
       return result.pruned
     },
     /** Begin pruning all members in the given time period */
     prune_members: (days: number) => {
-      if (days < 1) throw `The number of days must be 1 or more.`
-      // TODO: check if the bot has `KICK_MEMBERS` permission.
+      if (days < 1) throw new Error(Errors.PRUNE_MIN_DAYS)
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.KICK_MEMBERS]))
+        throw new Error(Errors.MISSING_KICK_MEMBERS)
       return client.discordRequestManager.post(endpoints.GUILD_PRUNE(data.id), { days })
     },
     // TODO: REQUEST THIS OVER WEBSOCKET WITH GET_GUILD_MEMBERS ENDPOINT
@@ -209,7 +220,9 @@ export const create_guild = (data: Create_Guild_Payload, client: Client) => {
     // },
     /** Returns the audit logs for the guild. Requires VIEW AUDIT LOGS permission */
     get_audit_logs: (options: Get_Audit_Logs_Options) => {
-      // TODO: check if the bot has VIEW_AUDIT_LOGS permission
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.VIEW_AUDIT_LOG]))
+        throw new Error(Errors.MISSING_VIEW_AUDIT_LOG)
+
       return client.discordRequestManager.get(endpoints.GUILD_AUDIT_LOGS(data.id), {
         ...options,
         limit: options.limit && options.limit >= 1 && options.limit <= 100 ? options.limit : 50
@@ -217,12 +230,14 @@ export const create_guild = (data: Create_Guild_Payload, client: Client) => {
     },
     /** Returns the guild embed object. Requires the MANAGE_GUILD permission. */
     get_embed: () => {
-      // TODO: check if the bot has the MANAGE_GUILD permission
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_GUILD]))
+        throw new Error(Errors.MISSING_MANAGE_GUILD)
       return client.discordRequestManager.get(endpoints.GUILD_EMBED(data.id))
     },
     /** Modify a guild embed object for the guild. Requires the MANAGE_GUILD permission. */
     edit_embed: (enabled: boolean, channel_id?: string | null) => {
-      // TODO: Requires the MANAGE_GUILD permission.
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_GUILD]))
+        throw new Error(Errors.MISSING_MANAGE_GUILD)
       return client.discordRequestManager.patch(endpoints.GUILD_EMBED(data.id), { enabled, channel_id })
     },
     /** Returns the code and uses of the vanity url for this server if it is enabled. Requires the MANAGE_GUILD permission. */
@@ -231,37 +246,44 @@ export const create_guild = (data: Create_Guild_Payload, client: Client) => {
     },
     /** Returns a list of integrations for the guild. Requires the MANAGE_GUILD permission. */
     get_integrations: () => {
-      // TODO: requires the MANAGE_GUILD permission
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_GUILD]))
+        throw new Error(Errors.MISSING_MANAGE_GUILD)
       return client.discordRequestManager.get(endpoints.GUILD_INTEGRATIONS(data.id))
     },
     /** Modify the behavior and settings of an integration object for the guild. Requires the MANAGE_GUILD permission. */
     edit_integration: (id: string, options: Edit_Integration_Options) => {
-      // TODO: requires the MANAGE_GUILD permission
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_GUILD]))
+        throw new Error(Errors.MISSING_MANAGE_GUILD)
       return client.discordRequestManager.patch(endpoints.GUILD_INTEGRATION(data.id, id), options)
     },
     /** Delete the attached integration object for the guild with this id. Requires MANAGE_GUILD permission. */
     delete_integration: (id: string) => {
-      // TODO: requires the MANAGE_GUILD permission
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_GUILD]))
+        throw new Error(Errors.MISSING_MANAGE_GUILD)
       return client.discordRequestManager.delete(endpoints.GUILD_INTEGRATION(data.id, id))
     },
     /** Sync an integration. Requires teh MANAGE_GUILD permission. */
     sync_integration: (id: string) => {
-      // TODO: requires MANAGE_GUILD
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_GUILD]))
+        throw new Error(Errors.MISSING_MANAGE_GUILD)
       return client.discordRequestManager.post(endpoints.GUILD_INTEGRATION_SYNC(data.id, id))
     },
     /** Returns a list of ban objects for the users banned from this guild. Requires the BAN_MEMBERS permission. */
     get_bans: () => {
-      // TODO: requires the BAN_MEMBERS permission
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.BAN_MEMBERS]))
+        throw new Error(Errors.MISSING_BAN_MEMBERS)
       return client.discordRequestManager.get(endpoints.GUILD_BANS(data.id))
     },
     /** Ban a user from the guild and optionally delete previous messages sent by the user. Requires teh BAN_MEMBERS permission. */
     ban: (id: string, options: BanOptions) => {
-      // TODO: requires the BAN_MEMBERS permission
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.BAN_MEMBERS]))
+        throw new Error(Errors.MISSING_BAN_MEMBERS)
       return client.discordRequestManager.put(endpoints.GUILD_BAN(data.id, id), options)
     },
     /** Remove the ban for a user. REquires BAN_MEMBERS permission */
     unban: (id: string) => {
-      // TODO: requires the BAN_MEMBERS permission
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.BAN_MEMBERS]))
+        throw new Error(Errors.MISSING_BAN_MEMBERS)
       return client.discordRequestManager.delete(endpoints.GUILD_BAN(data.id, id))
     },
     /** Check whether a member has certain permissions in this channel. */
@@ -297,12 +319,14 @@ export const create_guild = (data: Create_Guild_Payload, client: Client) => {
     },
     /** Modify a guilds settings. Requires the MANAGE_GUILD permission. */
     edit: (options: Guild_Edit_Options) => {
-      // TODO: requires the MANAGE_GUILD permission
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_GUILD]))
+        throw new Error(Errors.MISSING_MANAGE_GUILD)
       return client.discordRequestManager.patch(endpoints.GUILD(data.id), options)
     },
     /** Get all the invites for this guild. Requires MANAGE_GUILD permission */
     get_invites: () => {
-      // TODO: requires MANAGE_GUILD permission
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_GUILD]))
+        throw new Error(Errors.MISSING_MANAGE_GUILD)
       return client.discordRequestManager.get(endpoints.GUILD_INVITES(data.id))
     },
     /** Leave a guild */
@@ -315,7 +339,9 @@ export const create_guild = (data: Create_Guild_Payload, client: Client) => {
     },
     /** Returns a list of guild webhooks objects. Requires the MANAGE_WEBHOOKs permission. */
     get_webhooks: () => {
-      // TODO: requires MANAGE_WEBHOOKS
+      if (!bot_has_permission(data.id, client.bot_id, [Permissions.MANAGE_WEBHOOKS]))
+        throw new Error(Errors.MISSING_MANAGE_WEBHOOKS)
+
       return client.discordRequestManager.get(endpoints.GUILD_WEBHOOKS(data.id))
     }
   }
