+++ structures/message.ts
@@ -1,16 +1,19 @@
-import Client from '../module/client.ts'
-import { Message_Create_Options } from '../types/message.ts'
-import { endpoints } from '../constants/discord.ts'
-import { Channel_Types, MessageContent } from '../types/channel.ts'
-import { cache } from '../utils/cache.ts'
-import { create_user } from './user.ts'
-import { User_Payload } from '../types/guild.ts'
-import { Channel } from '../types/return-type.ts'
+import Client from "../module/client.ts"
+import { Message_Create_Options } from "../types/message.ts"
+import { endpoints } from "../constants/discord.ts"
+import { MessageContent } from "../types/channel.ts"
+import { cache } from "../utils/cache.ts"
+import { create_user } from "./user.ts"
+import { User_Payload } from "../types/guild.ts"
+import { Channel } from "../types/return-type.ts"
+import { bot_has_permission } from "../utils/permissions.ts"
+import { Errors } from "../types/errors.ts"
+import { Permissions } from "../types/permission.ts"
 
 export const create_message = (data: Message_Create_Options, client: Client) => {
   const base_message = {
     raw: () => data,
-    author: () => create_user({ ...data.author, avatar: data.author.avatar || '' }),
+    author: () => create_user({ ...data.author, avatar: data.author.avatar || "" }),
     id: () => data.id,
     type: () => data.type,
     timestamp: () => Date.parse(data.timestamp),
@@ -39,18 +42,22 @@ export const create_message = (data: Message_Create_Options, client: Client) => 
     channel: () => cache.channels.get(data.channel_id) as Channel,
 
     delete: (reason: string) => {
-      // TODO: Requires MANAGE_MESSAGES
-      if (data.author.id !== client.bot_id) {}
+      if (data.guild_id && !bot_has_permission(data.guild_id, client.bot_id, [Permissions.MANAGE_MESSAGES]))
+        throw new Error(Errors.MISSING_MANAGE_MESSAGES)
+      if (data.author.id !== client.bot_id) {
+      }
 
       client.discordRequestManager.delete(endpoints.CHANNEL_MESSAGE(data.channel_id, data.id), { reason })
     },
     /** Pin a message in a channel. Requires MANAGE_MESSAGES. Max pins allowed in a channel = 50. */
     pin: () => {
-      // TODO: Requires MANAGE_MESSAGES
+      if (data.guild_id && !bot_has_permission(data.guild_id, client.bot_id, [Permissions.MANAGE_MESSAGES]))
+        throw new Error(Errors.MISSING_MANAGE_MESSAGES)
       client.discordRequestManager.put(endpoints.CHANNEL_MESSAGE(data.channel_id, data.id))
     },
     unpin: () => {
-      // TODO: Requires MANAGE_MESSAGES
+      if (data.guild_id && !bot_has_permission(data.guild_id, client.bot_id, [Permissions.MANAGE_MESSAGES]))
+        throw new Error(Errors.MISSING_MANAGE_MESSAGES)
       client.discordRequestManager.delete(endpoints.CHANNEL_MESSAGE(data.channel_id, data.id))
     },
     /** Create a reaction for the message. Reaction takes the form of **name:id** for custom guild emoji, or Unicode characters. Requires READ_MESSAGE_HISTORY and ADD_REACTIONS */
@@ -63,32 +70,38 @@ export const create_message = (data: Message_Create_Options, client: Client) => 
     },
     /** Removes all reactions for all emojis on this message. */
     remove_all_reactions: () => {
-      // TODO: Requires MANAGE_MESSAGES
+      if (data.guild_id && !bot_has_permission(data.guild_id, client.bot_id, [Permissions.MANAGE_MESSAGES]))
+        throw new Error(Errors.MISSING_MANAGE_MESSAGES)
       client.discordRequestManager.delete(endpoints.CHANNEL_MESSAGE_REACTIONS(data.channel_id, data.id))
     },
     /** Removes all reactions for a single emoji on this message. Reaction takes the form of **name:id** for custom guild emoji, or Unicode characters. */
     remove_reaction_emoji: (reaction: string) => {
-      // TODO: Requires MANAGE_MESSAGES
+      if (data.guild_id && !bot_has_permission(data.guild_id, client.bot_id, [Permissions.MANAGE_MESSAGES]))
+        throw new Error(Errors.MISSING_MANAGE_MESSAGES)
       client.discordRequestManager.delete(endpoints.CHANNEL_MESSAGE_REACTION(data.channel_id, data.id, reaction))
     },
     /** Get a list of users that reacted with this emoji. */
     get_reactions: async (reaction: string) => {
-      const result = await client.discordRequestManager.get(endpoints.CHANNEL_MESSAGE_REACTION(data.channel_id, data.id, reaction)) as User_Payload[]
+      const result = (await client.discordRequestManager.get(
+        endpoints.CHANNEL_MESSAGE_REACTION(data.channel_id, data.id, reaction)
+      )) as User_Payload[]
       return result.map(res => create_user(res))
     },
     /** Edit the message. */
     edit: async (content: string | MessageContent) => {
-      if (data.author.id !== client.bot_id) throw 'You can only edit a message that was sent by the bot.'
-      if (data.type !== Channel_Types.DM) {
-        // TODO: check if the bot has SEND_MESSAGES permission
+      if (data.author.id !== client.bot_id) throw "You can only edit a message that was sent by the bot."
+
+      if (typeof content === "string") content = { content }
+
+      if (data.guild_id) {
+        if (!bot_has_permission(data.guild_id, client.bot_id, [Permissions.SEND_MESSAGES]))
+          throw new Error(Errors.MISSING_SEND_MESSAGES)
+
+        if (content.tts && !bot_has_permission(data.guild_id, client.bot_id, [Permissions.SEND_TTS_MESSAGES]))
+          throw new Error(Errors.MISSING_SEND_TTS_MESSAGE)
       }
 
-      if (typeof content === 'string') content = { content }
-      if (content.tts) {
-        // TODO: check if the bot has SEND_TTS_MESSAGE
-      }
-
-      // TODO: Check content length
+      if (content.content && content.content.length > 2000) throw new Error(Errors.MESSAGE_MAX_LENGTH)
 
       const result = await client.discordRequestManager.patch(endpoints.CHANNEL_MESSAGES(data.id), content)
       return create_message(result, client)
